se agrego policies y notifications

+++ app/Http/Controllers/FavoritoController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Auth;
 use App\Models\Favorito;
+use App\Models\User;
+use App\Notifications\NewFavorito;
+use App\Events\FavoritoEvent;
 
 class FavoritoController extends Controller
 {
@@ -14,15 +17,25 @@ class FavoritoController extends Controller
     }
 
     public function store(Request $request){
-        //return $request->favorito;
+        /*if ($request->user()->cannot('create', Favorito::class)) {
+            abort(403);
+        }*/
+        $this->authorize('create', Favorito::class);
         $favorito = Favorito::create([
             'titulo' => $request->favorito[1],
             'link' => 'https://hacker-news.firebaseio.com/v0/item/'.$request->favorito[0].'.json?print=pretty',
             'user_id' => Auth::user()->id,
         ]);
+        
+        /*User::all()->except($favorito->user_id)
+                    ->each(function(User $user) use ($favorito){
+                        $user->notify(new NewFavorito($favorito));
+                    });*/
+        event(new FavoritoEvent($favorito));
     }
 
     public function destroy(Favorito $favorito){
+        $this->authorize('delete', $favorito);
         $favorito->delete();
     }
 }
